pyproject.toml
[project]
name = "pdfdocx"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "docx-mailmerge2>=0.9.0",
    "fastexcel>=0.13.0",
    "jinja2>=3.1.6",
    "mistune>=3.1.3",
    "pendulum>=3.0.0",
    "pikepdf>=9.5.2",
    "polars>=1.23.0",
    "streamlit>=1.44.1",
    "typer>=0.15.1",
    "weasyprint>=65.0",
]

[project.optional-dependencies]
gui = [
    "marimo>=0.12.8",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
    "sphinx>=8.2.3",
    "sphinx-autobuild>=2024.10.3",
    "sphinx-rtd-theme>=3.0.2",
]
web = [
    "streamlit>=1.43.1",
]
